requirements.txt
Flask==2.2.3
Pillow==9.5.0
torch==1.13.1
yolov5==7.0.9
